Introduce ability to declare repositories in settings
This commit introduces an API in Settings which makes it possible
to declare repositories which would be used by default if a project
doesn't declare any explicit repository.

It's a cleaner replacement to the common:

allprojects {
   repositories {
     ...
   }
}

configuration block.

Settings repositories are used by any project which doesn't say
anything about repositories. They are also used for included
builds, meaning that if a build is included and that its projects
do not explicitly declare repositories, then they will be used
to resolve dependencies of included builds too.